# web/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ItemChanges": {
        "additionalProperties": false,
        "properties": {
          "note": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "title": "ItemChanges",
        "type": "object"
      },
      "ItemIn": {
        "additionalProperties": false,
        "properties": {
          "instrument_id": {
            "title": "Instrument Id",
            "type": "integer"
          },
          "note": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "required": [
          "instrument_id"
        ],
        "title": "ItemIn",
        "type": "object"
      },
      "ItemOut": {
        "properties": {
          "close": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Close"
          },
          "close_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Close Date"
          },
          "currency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "instrument_id": {
            "title": "Instrument Id",
            "type": "integer"
          },
          "isin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Isin"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "previous_close": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Close"
          },
          "stale": {
            "title": "Stale",
            "type": "boolean"
          },
          "ticker": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ticker"
          }
        },
        "required": [
          "id",
          "instrument_id",
          "name",
          "isin",
          "ticker",
          "currency",
          "note",
          "close",
          "close_date",
          "previous_close",
          "stale"
        ],
        "title": "ItemOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WatchlistIn": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "WatchlistIn",
        "type": "object"
      },
      "WatchlistOut": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ItemOut"
            },
            "title": "Items",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "items"
        ],
        "title": "WatchlistOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Folio",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/watchlists": {
      "get": {
        "operationId": "list_watchlists_api_v1_watchlists_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WatchlistOut"
                  },
                  "title": "Response List Watchlists Api V1 Watchlists Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Watchlists",
        "tags": [
          "watchlists"
        ]
      },
      "post": {
        "operationId": "create_watchlist_api_v1_watchlists_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WatchlistIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchlistOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Watchlist",
        "tags": [
          "watchlists"
        ]
      }
    },
    "/api/v1/watchlists/{watchlist_id}": {
      "delete": {
        "operationId": "delete_watchlist_api_v1_watchlists__watchlist_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "watchlist_id",
            "required": true,
            "schema": {
              "title": "Watchlist Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Watchlist",
        "tags": [
          "watchlists"
        ]
      },
      "patch": {
        "operationId": "rename_watchlist_api_v1_watchlists__watchlist_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "watchlist_id",
            "required": true,
            "schema": {
              "title": "Watchlist Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WatchlistIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchlistOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rename Watchlist",
        "tags": [
          "watchlists"
        ]
      }
    },
    "/api/v1/watchlists/{watchlist_id}/items": {
      "post": {
        "operationId": "add_item_api_v1_watchlists__watchlist_id__items_post",
        "parameters": [
          {
            "in": "path",
            "name": "watchlist_id",
            "required": true,
            "schema": {
              "title": "Watchlist Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchlistOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Item",
        "tags": [
          "watchlists"
        ]
      }
    },
    "/api/v1/watchlists/{watchlist_id}/items/{item_id}": {
      "delete": {
        "operationId": "remove_item_api_v1_watchlists__watchlist_id__items__item_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "watchlist_id",
            "required": true,
            "schema": {
              "title": "Watchlist Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchlistOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove Item",
        "tags": [
          "watchlists"
        ]
      },
      "patch": {
        "operationId": "edit_item_api_v1_watchlists__watchlist_id__items__item_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "watchlist_id",
            "required": true,
            "schema": {
              "title": "Watchlist Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemChanges"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchlistOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Edit Item",
        "tags": [
          "watchlists"
        ]
      }
    },
...
  }
}